FEAN-57 refactor: 패키지 수정
