+++ src/main/java/com/lotte/danuri/auth/dto/LoginReqDto.java
@@ -1,14 +1,14 @@
-package com.lotte.danuri.auth.jwt.dto;
+package com.lotte.danuri.auth.dto;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 @Getter
-@Builder
 @AllArgsConstructor
+@RequiredArgsConstructor
 public class LoginReqDto {
 
     @NotNull
